receitas/python/HackerNews/requirements.txt
ansi2html==1.5.2
Appium-Python-Client==1.0.2
attrs==20.2.0
docopt==0.6.2
docutils==0.16
gitignore==0.0.8
htmlmin==0.1.12
importlib-metadata==2.0.0
iniconfig==1.1.1
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-html==2.1.1
pytest-html-reporter==0.2.3
pytest-metadata==1.10.0
pytest-reporter==0.5.0
pytest-reporter-html1==0.8.1
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.25.11
zipp==3.4.0
